Expose node config subsystem interface
