+++ include/ccf/service/consensus_type.h
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the Apache 2.0 License.
+#pragma once
+
+enum ConsensusType
+{
+  CFT = 0,
+  BFT = 1
+};
