+++ include/ccf/service/reconfiguration_type.h
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the Apache 2.0 License.
+#pragma once
+
+enum ReconfigurationType
+{
+  ONE_TRANSACTION = 0,
+  TWO_TRANSACTION = 1
+};
