+++ include/ccf/service/service_config.h
@@ -1,0 +1,41 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the Apache 2.0 License.
+#pragma once
+
+#include "ccf/service/consensus_type.h"
+#include "ccf/service/reconfiguration_type.h"
+
+#include <cstdint>
+#include <optional>
+
+namespace ccf
+{
+
+  struct ServiceConfiguration
+  {
+    // Number of recovery shares required to decrypt the latest ledger secret
+    size_t recovery_threshold = 0;
+
+    ConsensusType consensus = ConsensusType::CFT;
+
+    /**
+     *  Fields below are added in 2.x
+     */
+
+    std::optional<size_t> maximum_node_certificate_validity_days = std::nullopt;
+    std::optional<size_t> maximum_service_certificate_validity_days =
+      std::nullopt;
+
+    std::optional<ReconfigurationType> reconfiguration_type = std::nullopt;
+
+    bool operator==(const ServiceConfiguration& other) const
+    {
+      return recovery_threshold == other.recovery_threshold &&
+        consensus == other.consensus &&
+        reconfiguration_type == other.reconfiguration_type &&
+        maximum_node_certificate_validity_days ==
+        other.maximum_node_certificate_validity_days;
+    }
+  };
+
+}
