+++ src/consensus/consensus_types.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "ccf/ds/unit_strings.h"
+#include "ccf/service/consensus_config.h"
 #include "ccf/service/tables/nodes.h"
 #include "ccf/tx_id.h"
 #include "enclave/consensus_type.h"
@@ -11,15 +12,6 @@
 
 namespace consensus
 {
-  struct Configuration
-  {
-    ConsensusType type = ConsensusType::CFT;
-    ds::TimeString message_timeout = {"100ms"};
-    ds::TimeString election_timeout = {"5000ms"};
-
-    bool operator==(const Configuration&) const = default;
-    bool operator!=(const Configuration&) const = default;
-  };
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(Configuration);
   DECLARE_JSON_REQUIRED_FIELDS(Configuration);
   DECLARE_JSON_OPTIONAL_FIELDS(
